reintroduce all dessins

+++ crates/dessins/src/chapter_1/polygon.rs
@@ -32,7 +32,7 @@ impl Polygon {
     }
 
     pub fn calculate_point(&self, i: u32) -> Point2 {
-        let angle = (2.0 * i as f32 * PI) / self.k.value as f32 + self.ad.value;
+        let angle = (2.0 * i as f32 * PI) / self.k.value + self.ad.value;
         let x = self.r.value * angle.cos();
         let y = self.r.value * angle.sin();
         pt2(x, y)
